B014_ollama-main/requirements.txt
pandas
streamlit
google-api-python-client
google-auth
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
langchain-ollama
# 추가로 필요한 의존성 (ChromaDB 구동용)
chromadb